third_party/WebKit/LayoutTests/imported/csswg-test/css-writing-modes-3/writing-mode-parsing-sideways-lr-001-expected.txt

This is a testharness.js-based test.
FAIL -webkit-writing-mode: sideways-lr assert_equals: expected "sideways-lr" but got "lr-tb"
Harness: the test ran to completion.
